@@ -1,12 +1,5 @@
-import XCTest
+import Testing
 @testable import KDL
 
-final class KDLTests: XCTestCase {
-    func testExample() throws {
-        // XCTest Documentation
-        // https://developer.apple.com/documentation/xctest
-
-        // Defining Test Cases and Test Methods
-        // https://developer.apple.com/documentation/xctest/defining_test_cases_and_test_methods
-    }
+struct KDLTests {
 }
